+++ common/src/main/java/net/spell_engine/api/spell/Spell.java
@@ -142,11 +142,13 @@ public class Spell {
         public Perks perks = new Perks();
         public static class Perks { Perks() { }
             public int ricochet = 0;
+            public int pierce = 0;
             // Watch out to add new fields to `copy` function too!
 
             public Perks copy() {
                 Perks copy = new Perks();
                 copy.ricochet = this.ricochet;
+                copy.pierce = this.pierce;
                 return copy;
             }
 
@@ -154,6 +156,7 @@ public class Spell {
             public String toString() {
                 return "Perks{" +
                         "ricochet=" + ricochet +
+                        ", pierce=" + pierce +
                         '}';
             }
         }