+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.9.29
 
+- Add universal Spell Projectile perks: ricochet, pierce
 - Add italian translation by Zano1999 #21
 - Fix HUD render blend glitch